requirements.txt
asgiref==3.5.2
autopep8==1.7.0
crispy-bootstrap5==0.7
Django==4.1
django-ajax-datatable==4.4.4
django-appconf==1.0.5
django-authtools==2.0.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-datatables-view==1.20.0
django-debug-toolbar==3.7.0
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-pivot==1.9.0
django-select2==8.0.0
django-tables2==2.4.1
et-xmlfile==1.1.0
gunicorn==20.1.0
jsonfield==3.1.0
MarkupPy==1.14
openpyxl==3.0.10
palettable==3.3.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.6
sqlparse==0.4.2
swapper==1.3.0
tablib==3.2.1
toml==0.10.2
tzdata==2022.2
whitenoise==6.2.0
